api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.1"
description = ""
authors = ["Mohamed Shegow <shegowmo@gmail.com>","Mattias Velamsson <smackdh@gmail.com>", "Suwon Baek <bkn020612@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2 = "2.9.9"
uWSGI = "^2.0.18,<2.1"
gunicorn = "21.2.0"  # Added this line
djangorestframework="3.14.0"
django-cors-headers="^4.2.0"
python-dotenv="1.0.0"
dj-rest-auth={version = "5.0.1", extras = ["with_social"]}
django-allauth="^0.55.0,<0.56.0"
djangorestframework-simplejwt="5.2.2"
pgvector="0.2.3"
moviepy="1.0.3"
openai="1.3.6"
tenacity="8.2.3"
pandas="2.0.3"
django-storages = { version = "1.14.2", extras = ["boto3"] }
boto3 = "^1.33.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
colorlog = "^6.7.0"
factory_boy="3.3.0"
faker="19.13.0"
black="23.10.1"

[tool.black]
line_length = 119

[tool.isort]
multi_line_output = 5
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
